requirements.txt
appdirs
astroid
attrs
black
Click
Flask
Flask-Cors
flask-jwt
Flask-MySQL
Flask-MySQLdb
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
mysqlclient
pathspec
PyJWT
pylint
PyMySQL
python-dotenv
regex
six
toml
typed-ast
werkzeug
wrapt
tensorflow
numpy
pillow